requirements.txt
# Web Scraping & HTTP
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
lxml==4.9.3
fake-useragent==1.4.0

# Database (Supabase)
supabase==2.0.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.23

# Data Processing
pandas==2.1.3
numpy==1.25.2
pillow==10.1.0

# Scheduling & Automation
schedule==1.2.0
apscheduler==3.10.4

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging & Monitoring
loguru==0.7.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
urllib3==2.1.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
